% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom-plots.R
\name{aggregateOverGenomeBins}
\alias{aggregateOverGenomeBins}
\title{Returns the distribution of query over a reference assembly
Given a query set of elements (a GRanges object) and a reference assembly (*e.g.
'hg38'), this will aggregate and count the distribution of the query elements
across bins of the reference genome}
\usage{
aggregateOverGenomeBins(query, refAssembly, binCount = 10000)
}
\arguments{
\item{query}{A GenomicRanges or GenomicRangesList object with query regions}

\item{refAssembly}{A character vector that will be used to grab chromosome
sizes with \code{getChromSizes}}

\item{binCount}{Number of bins to divide the chromosomes into}
}
\description{
Returns the distribution of query over a reference assembly
Given a query set of elements (a GRanges object) and a reference assembly (*e.g.
'hg38'), this will aggregate and count the distribution of the query elements
across bins of the reference genome
}
