% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{BSAggregate}
\alias{BSAggregate}
\title{BSaggregate -- Aggregate a BSDT across regions or region groups,
for multiple samples at a time.
This function is as BScombineByRegion, but can handle not only multiple
samples in BSDT, but also simultaneously multiple region sets by passing
a regionsGRL (GRangesList object).
you can use jExpr to do other functions.
Given a bisulfite data table as input, with an identifier column for
different samples; plus a GRanges objects with regions to aggregate.}
\usage{
BSAggregate(BSDT, regionsGRL, excludeGR = NULL, regionsGRL.length = NULL,
  splitFactor = NULL, keepCols = NULL, sumCols = NULL, jExpr = NULL,
  byRegionGroup = FALSE, keep.na = FALSE)
}
\arguments{
\item{BSDT}{The bisulfite data.table (output from one of the parsing
functions for methylation calls) that you wish to aggregate. It can
be a combined table, with individual samples identified by column passed
to splitFactor.}

\item{regionsGRL}{Regions across which you want to aggregate.}

\item{excludeGR}{A GenomicRanges object with regions you want to 
exclude from the aggregation function. These regions will be eliminated
from the input table and not counted.}

\item{jExpr}{You can pass a custom command in the j slot to data.table
specifying which columns to aggregate, and which functions to use. You
can use buildJ() to build a jExpr argument easily.}

\item{byRegionGroup}{You can aggregate by regionID or by regionGroupID; 
this reflects the regionsGRL that you pass; by default, BSAggregate will
aggregate each region individually -- scores will then be contiguous, and
the output is 1 row per region.
Turn on this flag to aggregate across all region groups, making the result
uncontiguous, and resulting in 1 row per *region group*.}
}
\description{
BSaggregate -- Aggregate a BSDT across regions or region groups,
for multiple samples at a time.
This function is as BScombineByRegion, but can handle not only multiple
samples in BSDT, but also simultaneously multiple region sets by passing
a regionsGRL (GRangesList object).
you can use jExpr to do other functions.
Given a bisulfite data table as input, with an identifier column for
different samples; plus a GRanges objects with regions to aggregate.
}
