% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{BSdtToGRanges}
\alias{BSdtToGRanges}
\title{Converts a list of data.tables (From BSreadbeds) into GRanges.}
\usage{
BSdtToGRanges(dtList)
}
\arguments{
\item{A}{list of data.tables}
}
\value{
A GRangesList object.
}
\description{
Converts a list of data.tables (From BSreadbeds) into GRanges.
}
