% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom-plots.R
\name{calcChromBins}
\alias{calcChromBins}
\title{Calculates the distribution of a query set over the genome}
\usage{
calcChromBins(query, bins)
}
\arguments{
\item{query}{A GenomicRanges or GenomicRangesList object with query regions}

\item{bins}{Pre-computed bins (as a GRangesList object) to aggregate
over; for example, these could be genome bins}
}
\value{
A data.table showing where on which chromosomes ranges are distributed.
}
\description{
Returns a data.table showing counts of regions in GR, in the bins
In other words, where on which chromosomes are the ranges distributed?
You must provide binned regions.
}
