% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qthist.R
\name{calcDivisions}
\alias{calcDivisions}
\title{Internal helper function for \code{plotQTHist}}
\usage{
calcDivisions(widths, bins = NULL, quantile = NULL)
}
\description{
If the bins or quantiles for the hist. are specified by the user, those are used
otherwise, calculates bins based on size of the dataset, and quantile based on bins
Returns the divisions that will be used in plotting the histogram
}
