% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-overlap-plots.R
\name{calcExpectedDistribution}
\alias{calcExpectedDistribution}
\title{Calculate the expected fraction of reads in features.}
\usage{
calcExpectedDistribution(frif, genome_size, feature_names = NA)
}
\arguments{
\item{frif}{Results from \code{calcFRiF}}

\item{genome_size}{Numeric value representing the size of the genome in
bases.}

\item{feature_names}{An optional character vector of feature names, in the 
same order as the GenomicRanges or GenomicRangesList 
object.}
}
\description{
Calculate the expected fraction of reads in features.
}
