% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-plots.R
\name{calcExpectedPropPartitions}
\alias{calcExpectedPropPartitions}
\title{Calculates expected partition overlap based on base pair contribution
of each feature (partition) to genome .Expected and observed overlaps
are then compared.}
\usage{
calcExpectedPropPartitions(
  query,
  partitionList,
  genomeSize = NULL,
  remainder = "intergenic"
)
}
\arguments{
\item{query}{GRanges or GRangesList with regions to classify.}

\item{partitionList}{NAMED list of genomic partitions
GRanges.}

\item{genomeSize}{The number of bases in the query genome. In other words, 
the sum of all chromosome sizes.}

\item{remainder}{Which partition do you want to account for 'everything 
else'?}
}
\value{
A data.frame assigning each element of a GRanges object to a
    partition from a previously provided partitionList.
}
\description{
Calculates expected partition overlap based on base pair contribution
of each feature (partition) to genome .Expected and observed overlaps
are then compared.
}
\examples{
partitionList = genomePartitionList(geneModels_hg19$genesGR,
                                    geneModels_hg19$exonsGR,
                                    geneModels_hg19$threeUTRGR, 
                                    geneModels_hg19$fiveUTRGR)
chromSizes = getChromSizes('hg19')
genomeSize = sum(chromSizes)
calcExpectedPropPartitions(vistaEnhancers, partitionList, genomeSize)
}
