% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-distance-plots.R,
%   R/feature-overlap-plots.R
\name{calcFRiF}
\alias{calcFRiF}
\title{Calculate the Fraction of Regions in Features (FRiF)}
\usage{
calcFRiF(query, total)

calcFRiF(query, total)
}
\arguments{
\item{query}{A GenomicRanges object with query regions and a metadata
column named 'counts' which represents the number of reads
or bases that aligned to the corresponding regions.}

\item{total}{A numeric value presenting the total number of aligned 
reads/bases.}
}
\description{
This function calculates the fraction of regions in a feature set and
returns the cumulative sum of reads, cumulative size of covered features, 
the fraction of reads in those features, and the number of total features.

This function calculates the fraction of regions in a feature set and
returns the cumulative sum of reads, cumulative size of covered features, 
the fraction of reads in those features, and the number of total features.
}
\examples{
calcFRiF()
}
\keyword{FRiF}
