% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest-genes.R
\name{calcNearestGenes}
\alias{calcNearestGenes}
\title{Given a query and set of annotations, this function will calculate
the nearest annotation to each region in the region set, as well
as the nearest gene type and the distance to the nearest gene.}
\usage{
calcNearestGenes(
  query,
  annotations,
  gene_name_key = "gene_id",
  gene_type_key = "gene_biotype"
)
}
\arguments{
\item{query}{A GRanges or GRangesList object with query sets}

\item{annotations}{A GRanges or GRangesList object with annotation sets}
}
\value{
A data table that contains observations for each genomic region
        and the associated aforementioned annotations.
}
\description{
Given a query and set of annotations, this function will calculate
the nearest annotation to each region in the region set, as well
as the nearest gene type and the distance to the nearest gene.
}
\examples{
queryFile = system.file("extdata", "vistaEnhancers.bed.gz", package="GenomicDistributions")
query = rtracklayer::import(queryFile)
data(TSS_hg19)

queryAnnotated = calcNearestGenes(query, TSS_hg19)
}
