% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor-distances.R
\name{calcNearestNeighbors}
\alias{calcNearestNeighbors}
\title{Group regions from the same chromosome together and
compute the distance of a region to its nearest neighbor. 
Distances are then lumped into a numeric vector.}
\usage{
calcNearestNeighbors(query, correctRef = "None")
}
\arguments{
\item{query}{A GRanges or GRangesList object.}

\item{correctRef}{A string indicating the reference genome
to use if Nearest neighbor distances are corrected for the 
number of regions in a regionSet.}
}
\value{
A numeric vector or list of vectors containing the
 distance of regions to their nearest neighbors.
}
\description{
Group regions from the same chromosome together and
compute the distance of a region to its nearest neighbor. 
Distances are then lumped into a numeric vector.
}
\examples{
Nneighbors = calcNearestNeighbors(vistaEnhancers)
}
