% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openRegionsCalcAndPlot.R
\name{calcOpenSignal}
\alias{calcOpenSignal}
\title{The function calcOpenSignal takes the input BED file, overlaps it with all defined
open chromatin regions across cell types and returns a matrix, where each row is 
the input genomic region (if overlap was found), each column is a cell type and the 
value is a normalized ATAC-seq signal in the region for the cell type.}
\usage{
calcOpenSignal(
  bedInput,
  genome = "hg19",
  useSimpleCache = T,
  cacheDir = tempdir()
)
}
\arguments{
\item{bedInput}{Genomic regions to be analyzed in form of data.table, data.frame, or 
GRanges object}

\item{genome}{- genome version can be either hg19 (default) or hg38}

\item{cacheDir}{- set up a directory for simpleCache, default is a tempdir}

\item{useSimpleCahce}{- defines if simpleCache should be used to upload signal matrix, 
defaults to T}
}
\description{
The function calcOpenSignal takes the input BED file, overlaps it with all defined
open chromatin regions across cell types and returns a matrix, where each row is 
the input genomic region (if overlap was found), each column is a cell type and the 
value is a normalized ATAC-seq signal in the region for the cell type.
}
\examples{
\dontrun{
signalMatrix = calcOpenSignal(bedInput)
}
}
