% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-plots.R
\name{calcPartitionPercents}
\alias{calcPartitionPercents}
\title{Calculate the percentage overlap of a list of GRanges object against a 
list of known partitions.}
\usage{
calcPartitionPercents(listGR, partitionList, backgroundGR = NULL)
}
\arguments{
\item{listGR}{A list of GRanges objects. Produced by 
\code{calcPartitionPercents}}

\item{partitionList}{A list with GRanges objects corresponding to
the different partitions.}

\item{backgroundGR}{A GRanges object to remove from listGR overlap 
as background.}
}
\value{
A named list of the frequency of input regions against provided
    partitions.
}
\description{
Calculate the percentage overlap of a list of GRanges object against a 
list of known partitions.
}
\examples{
f = system.file("extdata", "vistaEnhancers.bed.gz", 
    package="GenomicDistributions")
query = rtracklayer::import(f)
geneModels = getGeneModels("hg38")
partitionList = genomePartitionList(geneModels$genesGR, geneModels$exonsGR)
partPerc = calcPartitionPercents(query, partitionList)
}
