% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-plots.R
\name{calcPropPartitions}
\alias{calcPropPartitions}
\title{Calculates the distribution of overlaps (in base pairs) between 
query and arbitrary genomic partitions.}
\usage{
calcPropPartitions(query, partitionList, remainder = "intergenic")
}
\arguments{
\item{query}{GRanges or GRangesList with regions to classify}

\item{partitionList}{NAMED list of genomic partitions
GRanges.}

\item{remainder}{A character vector to assign any query regions that do
not overlap with anything in the partitionList. Defaults to "intergenic"}
}
\value{
A data.frame giving total overlap [bp] of a GRanges object to a
    partition from a previously provided partitionList.
}
\description{
Takes overlaps between query and a partition. The output is the number
of base pairs that fall into overlap. Frequency then gives the proportions
of query that overlap with a given partition
}
\examples{
partitionList = genomePartitionList(geneModels_hg19$genesGR, 
                                    geneModels_hg19$exonsGR,
                                    geneModels_hg19$threeUTRGR, 
                                    geneModels_hg19$fiveUTRGR)
calcPropPartitions(vistaEnhancers, partitionList)
}
