% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qthist.R
\name{calcWidth}
\alias{calcWidth}
\title{Calculate the widths of regions}
\usage{
calcWidth(query)
}
\arguments{
\item{query}{A GRanges or GRangesList object with query sets}
}
\value{
A vector of the widths (end-start coordinates) of GRanges objects.
}
\description{
The length of a genomic region (the distance between the start and end) 
is called the width
When given a query set of genomic regions, this function returns the width
}
\examples{
regWidths = calcWidth(vistaEnhancers)
}
