% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{chromSizes}
\alias{chromSizes}
\title{Chromosome sizes}
\format{A list of vectors with one vector per genome assembly}
\source{
BSGenome package
}
\description{
A dataset containing chromosome sizes for common reference genome assemblies
}
