% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{chromSizes_hg19}
\alias{chromSizes_hg19}
\title{hg19 chromosome sizes}
\format{
A named vectors of lengths with one item per chromosome
}
\source{
BSgenome.Hsapiens.UCSC.hg19 package
}
\description{
A dataset containing chromosome sizes for Homo Sapiens hg38 genome assembly
}
