% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{chromSizes_mm10}
\alias{chromSizes_mm10}
\title{Chromosome sizes}
\format{A named vectors of lengths with one item per chromosome}
\source{
BSGenome package
}
\description{
A dataset containing chromosome sizes for common reference genome assemblies
}
