% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.requireAndReturn}
\alias{.requireAndReturn}
\title{Checks to make sure a package object is installed,
and if so, returns it. If the library is not installed, it issues a warning
and returns NULL.}
\usage{
.requireAndReturn(BSgenomeString)
}
\arguments{
\item{BSgenomeString}{A BSgenome compatible genome string.}
}
\value{
A BSgenome object if installed.
}
\description{
Checks to make sure a package object is installed,
and if so, returns it. If the library is not installed, it issues a warning
and returns NULL.
}
