% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.validateInputs}
\alias{.validateInputs}
\title{Checks class of the list of variables. To be used in functions}
\usage{
.validateInputs(checkList)
}
\arguments{
\item{checkList}{list of object to check, e.g. 
list(varname=c("data.frame", "numeric")). 
Multiuple strings in the vector are treated as OR.}
}
\description{
Checks class of the list of variables. To be used in functions
}
\examples{
x <- function(var1) {
    cl = list(var1=c("numeric","character"))
    .validateInputs(cl)
    return(var1^2)
}
}
