% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dtToGr}
\alias{dtToGr}
\title{Converts a data.table (DT) object to a GenomicRanges (GR) object. Tries to be
intelligent, guessing chr and start, but you have to supply end or other
columns if you want them to be carried into the GR.}
\usage{
dtToGr(DT, chr = "chr", start = "start", end = NA, strand = NA,
  name = NA, splitFactor = NA, metaCols = NA)
}
\description{
Converts a data.table (DT) object to a GenomicRanges (GR) object. Tries to be
intelligent, guessing chr and start, but you have to supply end or other
columns if you want them to be carried into the GR.
}
