% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{exampleBED}
\alias{exampleBED}
\title{exampleBED}
\format{data.frame object}
\usage{
data(exampleBED)
}
\description{
A dataset containing example hg19 regions. The dataset is an example of H3K27ac 
regions in PBMCs.
}
