% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{exampleopenSignalMatrix_hg19}
\alias{exampleopenSignalMatrix_hg19}
\title{A dataset containing a subset of open chromatin regions across all cell types defined by ENCODE for Homo Sapiens hg19}
\format{
data.frame, rows represent whole selection of open 
chromatin regions across all cell types defined by ENCODE, columns are 
individual cell types and values are normalized open chromatin signal values.
}
\source{
\url{http://big.databio.org/open_chromatin_matrix/openSignalMatrix_hg19_quantileNormalized_round4.txt.gz}
}
\description{
Preparation steps:
\enumerate{
   \item{made a universe of regions by merging regions across cell types defined as opened in ENCODE}
   \item{took bigwig files from ENCODE for individual cell types, merged replicates, filtered out blacklisted sites}
   \item{evaluated the signal above regions defined by previous step}
   \item{performed quantile normalization}
   \item{subsetted it}
}
}
