% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom-plots.R
\name{featureAggregateDistribution}
\alias{featureAggregateDistribution}
\title{Calculates the distribution of a query set over the genome}
\usage{
featureAggregateDistribution(query, features)
}
\arguments{
\item{query}{A GenomicRanges or GenomicRangesList object with query regions}

\item{features}{Pre-computed features (as a GRangesList object) to aggregate
over; for example, these could be genome bins}
}
\description{
Returns a data.table showing counts of regions in GR, in the bins
In other words, where on which chromosomes are the ranges distributed?
You must provide *either* a refAssembly (and binCount), or a GRangesList object
as already binned regions
}
