% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{geneModels_mm10}
\alias{geneModels_mm10}
\alias{geneModels_mm10}
\alias{geneModels_mm10}
\title{Gene models}
\format{A named vectors of lengths with one item per chromosome}
\source{
BSGenome package

BSGenome package

BSGenome package
}
\description{
A dataset containing chromosome sizes for common reference genome assemblies

A dataset containing chromosome sizes for common reference genome assemblies

A dataset containing chromosome sizes for common reference genome assemblies
}
