% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-plots.R
\name{genomeLabel}
\alias{genomeLabel}
\title{Converts a nucleotide count into a label with abbreviation}
\usage{
genomeLabel(x)
}
\arguments{
\item{x}{base count}
}
\value{
A label with 'kb' or 'mb' appended if appropriate
}
\description{
Converts a nucleotide count into a label with abbreviation
}
