% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-plots.R
\name{genomePartitionList}
\alias{genomePartitionList}
\title{Create a basic genome partition list of genes, exons, introns, and intergenic}
\usage{
genomePartitionList(genesGR, exonsGR)
}
\arguments{
\item{genesGR}{a GRanges object of gene coordinates}

\item{exonsGR}{a GRanges object of exons coordinates}
}
\description{
Given GRanges for genes, and a GRanges for exons, returns a list of GRanges
corresponding to various breakdown of the genome, based on the given
annotations; it gives you proximal and core promoters, exons, and introns. To
be used as a partionList for assignPartitions()
}
\examples{
eload(loadGencodeGenes("human"))
partitionList = genomePartitionList(SV$gencodeContainer$genesGR, SV$gencodeContainer$exonsGR)
}
