% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom-plots.R
\name{genomicDistribution}
\alias{genomicDistribution}
\title{Calculates the distribution of a query set over the genome}
\usage{
genomicDistribution(query, genome, binCount = 1000, genomeBins = NULL)
}
\arguments{
\item{query}{A GenomicRanges or GenomicRangesList object with query regions}

\item{genome}{A character vector that will be used to grab a BSGenome object
by \code{binBSGenome}}
}
\description{
Returns a data.table showing counts of regions in GR, in the bins
In other words, where on which chromosomes are the ranges distributed?
}
