% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{GenomicDistributions-package}
\alias{GenomicDistributions}
\alias{GenomicDistributions-package}
\title{Produces summaries and plots of features distributed across genomes}
\description{
If you have a set of genomic ranges, the GenomicDistributions R package can
help you with some simple visualizations. Currently, it can produce two kinds
of plots: First, the chromosome distribution plot, which visualizes how your
regions are distributed over chromosomes; and second, the feature
distribution plot, which visualizes how your regions are distributed relative
to a feature of interest, like Transcription Start Sites (TSSs).
}
\seealso{
Useful links:
\itemize{
  \item \url{http://code.databio.org/GenomicDistributions}
  \item Report bugs at \url{http://github.com/databio/GenomicDistributions}
}

}
\author{
\strong{Maintainer}: Kristyna Kupkova \email{kristynakupkova@gmail.com}

Authors:
\itemize{
  \item Jose Verdezoto
  \item Tessa Danehy
  \item John Lawson
  \item Jose Verdezoto
  \item Michal Stolarczyk
  \item Jason Smith
  \item Bingjie Xue
  \item Sophia Rogers
  \item John Stubbs
  \item Nathan C. Sheffield \email{nathan@code.databio.org}
}

}
