% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{GenomicDistributions}
\alias{GenomicDistributions}
\alias{GenomicDistributions-package}
\title{Provides useful genome functions}
\description{
Provides useful genome functions
}
\references{
\url{http://github.com/nsheff}
}
\author{
Nathan C. Sheffield
}
