% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-plots.R
\name{genomicPartitions}
\alias{genomicPartitions}
\title{Calculates the distribution of overlaps for a query set to genomic partitions}
\usage{
genomicPartitions(query, refAssembly)
}
\arguments{
\item{query}{A GenomicRanges or GenomicRangesList object with query regions}

\item{refAssembly}{A character vector specifying the reference genome
assembly (*e.g.* 'hg19'). This will be used to grab chromosome sizes with
\code{getTSSs}.}
}
\description{
Calculates the distribution of overlaps for a query set to genomic partitions
}
