% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-distance-plots.R
\name{getExpectedFeatures}
\alias{getExpectedFeatures}
\title{Internal helper function for \code{plotFRiF}}
\usage{
getExpectedFeatures(query, genome_size)
}
\arguments{
\item{query}{A GRanges object.}

\item{genome_size}{Numeric value representing the size of the genome in
bases.}
}
\description{
Internal helper function for \code{plotFRiF}
}
