% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{getGeneModels}
\alias{getGeneModels}
\title{Returns built-in gene models for a given reference assembly}
\usage{
getGeneModels(refAssembly)
}
\arguments{
\item{refAssembly}{A string identifier for the reference assembly}
}
\value{
A list containing the gene models corresponding to a
specific reference assembly.
}
\description{
Some functions require gene models, which can obtained from any source.
This function allows you to retrieve a few common built-in ones.
}
\examples{
getGeneModels("hg19")
}
