% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{getReferenceData}
\alias{getReferenceData}
\title{Get reference data for a specified assembly}
\usage{
getReferenceData(refAssembly, tagline)
}
\arguments{
\item{refAssembly}{Reference assembly string (e.g. 'hg38')}

\item{tagline}{The string that was used to identify data of a given type in 
the data building step. It's used for the filename so we know
what to load, and is what makes this function generic (so it 
can load different data types).}
}
\value{
A requested and included package data object.
}
\description{
This is a generic getter function that will return a data object requested,
if it is included in the built-in data with the GenomicDistributions package 
or GenomicDistributionsData package (if installed). Data objects can 
be requested for different reference assemblies and data types (specified by
a tagline, which is a unique string identifying the data type).
}
