% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildReferenceData.R
\name{getTssFromGTF}
\alias{getTssFromGTF}
\title{Get transcription start sites (TSSs) from a remote or local GTF file}
\usage{
getTssFromGTF(source, destDir, convertEnsemblUCSC)
}
\arguments{
\item{source}{a string that is either a path to a local or remote GTF}

\item{destDir}{a string that indicates the path to the directory where the downloaded GTF file should be stored}

\item{convertEnsemblUCSC}{a logical indicating whether Ensembl style chromosome annotation should be changed to UCSC style}
}
\value{
a list of GRanges objects
}
\description{
Get transcription start sites (TSSs) from a remote or local GTF file
}
\examples{
CElegansGtfUrl = "http://ftp.ensembl.org/pub/release-103/gtf/caenorhabditis_elegans/Caenorhabditis_elegans.WBcel235.103.gtf.gz"
CElegansTss = getTssFromGTF(CElegansGtfUrl, getwd(), TRUE)
}
