% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{grToDt}
\alias{grToDt}
\title{Convert a GenomicRanges into a data.table.}
\usage{
grToDt(GR)
}
\arguments{
\item{A}{Granges object}
}
\value{
A data.table object.
}
\description{
Convert a GenomicRanges into a data.table.
}
