% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadEnsDb}
\alias{loadEnsDb}
\title{Loads Ensemble gene annotations (EnsDb objects) from UCSC-style character vectors}
\usage{
loadEnsDb(genomeBuild)
}
\arguments{
\item{genomeBuild}{One of 'hg19', 'hg38', 'mm10', 'mm9', or 'grch38'}
}
\description{
This function will let you use a simple character vector (e.g. 'hg19') to
load and then return BSgenome objects. This lets you avoid having to use the
more complex annotation for a complete BSgenome object (e.g.
BSgenome.Hsapiens.UCSC.hg38.masked)
}
\examples{
\dontrun{
ensDb = loadEnsDb('hg19')
}
}
