% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor-distances.R
\name{neighbordt}
\alias{neighbordt}
\title{Internal helper function to calculate distance 
between neighboring regions.}
\usage{
neighbordt(querydt)
}
\arguments{
\item{querydt}{A data table with regions grouped according to
chromosome.}
}
\value{
A numeric vector with the distances in bp
}
\description{
Internal helper function to calculate distance 
between neighboring regions.
}
