% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-plots.R
\name{partitionPlot}
\alias{partitionPlot}
\title{Produces a barplot showing how query regions of interest are distributed
across a given partition list}
\usage{
partitionPlot(assignedPartitions, labels = NULL)
}
\arguments{
\item{listGR}{a list of GRanges objects with independent region sets of interest}

\item{partitionList}{A priority-ordered and named list of GRanges objects to test
query regions for overlaps}
}
\description{
This function can be used to test a GRanges object against any arbitrary list
of genome partitions. The partition list is a priority-ordered list of
GRanges objects. Each region in the query will be assigned to a given
partition that it overlaps with the highest priority.
}
