% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-plots.R
\name{percentReshape}
\alias{percentReshape}
\title{Internal helper function}
\usage{
percentReshape(percentList, percentdf)
}
\arguments{
\item{percentList}{A list with either percentages or background
corrected values}

\item{percentdf}{A character vector pointing to the appropriate
data frame inside percentList}
}
\value{
A ggplot object
}
\description{
Used for \code{plotPartitionPercents}
}
