% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-overlap-plots.R
\name{plotcFRiF}
\alias{plotcFRiF}
\title{Plot cumulative Fraction of Reads in Features (cFRiF)}
\usage{
plotcFRiF(frif, feature_names = NA)
}
\arguments{
\item{frif}{Results from \code{calcFRiF}}

\item{feature_names}{An optional character vector of feature names, in the 
same order as the GenomicRanges or GenomicRangesList 
object.}
}
\description{
This function plots the cumulative fraction of reads in a set of features
}
\keyword{cFRiF}
