% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-plots.R
\name{plotCumulativePartitions}
\alias{plotCumulativePartitions}
\title{Plot the cumulative distribution of regions in features}
\usage{
plotCumulativePartitions(assignedPartitions, feature_names = NULL)
}
\arguments{
\item{assignedPartitions}{Results from \code{calcCumulativePartitions}}

\item{feature_names}{An optional character vector of feature names, in the 
same order as the GenomicRanges or GenomicRangesList object.}
}
\value{
A ggplot object of the cumulative distribution of regions in 
    features.
}
\description{
This function plots the cumulative distribution of regions across a 
feature set.
}
\examples{
p = calcCumulativePartitionsRef(vistaEnhancers, "hg19")
cumuPlot = plotCumulativePartitions(p)
}
