% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content-plots.R
\name{plotDinuclFreq}
\alias{plotDinuclFreq}
\title{Plot dinuclotide content over genomic ranges}
\usage{
plotDinuclFreq(DNFDataTable)
}
\arguments{
\item{DNFDataTable}{A data.table of dinucleotide counts}
}
\value{
A ggplot object plotting distribution of dinucleotide content per dinucleotide
}
\description{
Given \code{calcDinuclFreq} results, this function 
generates a violin plot of dinucleotide frequency
}
\examples{

DNFDataTable = data.table::data.table(GC = rnorm(400, mean=0.5, sd=0.1), 
CG = rnorm(400, mean=0.5, sd=0.5), 
AT = rnorm(400, mean=0.5, sd=1), 
TA = rnorm(400, mean=0.5, sd=1.5))

DNFPlot =  plotDinuclFreq(DNFDataTable)
}
