% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-overlap-plots.R
\name{plotFRiF}
\alias{plotFRiF}
\title{Plot Fraction of Reads in Features (FRiF)}
\usage{
plotFRiF(dists, feature_names = NA)
}
\arguments{
\item{dists}{Results from \code{calcExpectedDistribution}}

\item{feature_names}{An optional character vector of feature names, in the 
same order as the dists object.}
}
\description{
This function plots the log10 of the observed/expected distribution of
reads/bases in a set of features.
}
\keyword{FRiF}
