% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom-plots.R
\name{plotGenomicDist}
\alias{plotGenomicDist}
\title{Plot distribution over chromosomes
Plots result from \code{genomicDistribution} calculation}
\usage{
plotGenomicDist(GD, binCount = 1000,
  plotTitle = "Distribution over chromosomes")
}
\arguments{
\item{GD}{The output from the genomicDistribution function}

\item{binCount}{Number of bins (should match the call to
\code{genomicDistribution}}
}
\description{
Plot distribution over chromosomes
Plots result from \code{genomicDistribution} calculation
}
