% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-plots.R
\name{plotPartitionPercents}
\alias{plotPartitionPercents}
\title{Plot the percentage overlap of GRanges objects against defined partitions.}
\usage{
plotPartitionPercents(percentList)
}
\arguments{
\item{percentList}{A named list of percentage overlap of regions to genomic
partitions. Produced by \code{calcPartitionPercents}}
}
\value{
A ggplot object using a barplot to show either percentage overlap
        or genomic background corrected overlap data
}
\description{
Plot the percentage overlap of GRanges objects against defined partitions.
}
\examples{
partitionList = genomePartitionList(geneModels_hg19$genesGR, 
                                    geneModels_hg19$exonsGR,
                                    geneModels_hg19$threeUTRGR, 
                                    geneModels_hg19$fiveUTRGR)
partPerc = calcPartitionPercents(vistaEnhancers, partitionList)
p = plotPartitionPercents(partPerc)
}
