% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-plots.R
\name{plotPartitions}
\alias{plotPartitions}
\title{Produces a barplot showing how query regions of interest are distributed
across a given partition list}
\usage{
plotPartitions(assignedPartitions, labels = NULL)
}
\arguments{
\item{assignedPartitions}{A table holding the frequency of assignment to
each of the partitions. Produced by \code{calcPartitions}}

\item{labels}{Character vector with labels for the partitions (optional). By
default it will use the names from the first argument.}
}
\value{
A ggplot object using a barplot to show the distribution of the query 
 regions across a given partition list.
}
\description{
This function can be used to test a GRanges object against any arbitrary 
list of genome partitions. The partition list is a priority-ordered list of
GRanges objects. Each region in the query will be assigned to a given
partition that it overlaps with the highest priority.
}
\examples{
f = system.file("extdata", "vistaEnhancers.bed.gz",
    package="GenomicDistributions")
query = rtracklayer::import(f)
p = calcPartitionsRef(query, "hg19")
partPlot = plotPartitions(p)
}
