% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qthist.R
\name{plotQTHistIndep}
\alias{plotQTHistIndep}
\title{Plot quantile-trimmed histogram independently}
\usage{
plotQTHistIndep(
  widths,
  EndBarColor = "gray57",
  MiddleBarColor = "gray27",
  quantile = NULL,
  bins = NULL
)
}
\arguments{
\item{widths}{Results from \code{calcWidths}}

\item{EndBarColor}{Color for the quantile bars on both ends of the graph (optional)}

\item{MiddleBarColor}{Color for the bars in the middle of the graph (optional)}

\item{quantile}{Quantile of data to be contained in each end bar (optional)
Quantiles must be under .2, optimal size is under .1}

\item{bins}{The number of bins for the histogram to allocate data to. (optional)}
}
\description{
Given the results from \code{calcWidth}, plots a histogram of widths
}
\details{
x-axis breaks for the frequency calculations are based on the "divisions" results from 
helper function \code{calcDivisions}
The difference between  plotQTHistIndep and plotQTHist is that the 'Indep'
version will return a list of plots that have had their bins calculated
independently; the normal version will plot them on the same x and y axis.
}
\examples{
plotQTHistIndep(runif(500)*1000)
}
