% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildReferenceData.R
\name{retrieveFile}
\alias{retrieveFile}
\title{Read local or remote file}
\usage{
retrieveFile(source, destDir = NULL)
}
\arguments{
\item{source}{a string that is either a path to a local or remote GTF}

\item{destDir}{a string that indicates the path to the directory where the downloaded GTF file should be stored}
}
\value{
data.frame retrieved file path
}
\description{
Read local or remote file
}
\examples{
CElegansGtfUrl = "http://ftp.ensembl.org/pub/release-103/gtf/caenorhabditis_elegans/Caenorhabditis_elegans.WBcel235.103.gtf.gz"
CElegansGtf = retrieveFile(CElegansGtfUrl)
}
