% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{setB_100}
\alias{setB_100}
\title{Example BED file read with rtracklayer::import}
\format{
GenomicRanges::GRanges
}
\usage{
data(setB_100)
}
\description{
Example BED file read with rtracklayer::import
}
\keyword{datasets}
