% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-plots.R
\name{setLabels}
\alias{setLabels}
\title{Internal helper function for \code{plotcFRiF} and \code{plotFRiF}}
\usage{
setLabels(query)
}
\arguments{
\item{query}{A GRanges object}
}
\description{
Internal helper function for \code{plotcFRiF} and \code{plotFRiF}
}
