% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{theme_blank_facet_label}
\alias{theme_blank_facet_label}
\title{Clear ggplot face label
Usually ggplot2 facets are labeled with boxes surrounding the label. This
function removes the box, so it's a simple label for each facet.}
\usage{
theme_blank_facet_label()
}
\description{
Clear ggplot face label
Usually ggplot2 facets are labeled with boxes surrounding the label. This
function removes the box, so it's a simple label for each facet.
}
