% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{TSS_mm9}
\alias{TSS_mm9}
\title{TSS locations}
\format{A named vectors of lengths with one item per chromosome}
\source{
BSGenome package
}
\description{
A dataset containing chromosome sizes for common reference genome assemblies
}
