% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-plots.R
\name{TSSDistance}
\alias{TSSDistance}
\title{Calculates the distribution of distances from a query set to closest TSS}
\usage{
TSSDistance(query, refAssembly)
}
\arguments{
\item{query}{A GenomicRanges or GenomicRangesList object with query regions}

\item{refAssembly}{A character vector specifying the reference genome
assembly (*e.g.* 'hg19'). THis will be used to grab chromosome sizes with
\code{getTSSs}.}
}
\description{
Given a query GRanges object and an assembly string, this function will grab the
TSS list for the given reference assembly and then calculate the distance from
each query feature to the closest TSS.
}
