% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vistaEnhancers}
\alias{vistaEnhancers}
\title{Example BED file read with rtracklayer::import}
\format{
GenomicRanges::GRanges
}
\usage{
data(vistaEnhancers)
}
\description{
Example BED file read with rtracklayer::import
}
\keyword{datasets}
