% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom-plots.R
\name{binChroms}
\alias{binChroms}
\title{Splits a chromosome into binSize
Given a list of chromosomes with corresponding sizes, this script will produce
(roughly) evenly-sized bins across the chromosomes}
\usage{
binChroms(binCount, chromSizes)
}
\arguments{
\item{binCount}{number of bins per chromosome}

\item{chromSizes}{a named list of size (length) for each chromosome}
}
\description{
Splits a chromosome into binSize
Given a list of chromosomes with corresponding sizes, this script will produce
(roughly) evenly-sized bins across the chromosomes
}
