% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest-gene.R
\name{calcNearestGenes}
\alias{calcNearestGenes}
\title{Find overlapping annotations for a given query and annotation
set. Annotate the query based on the distance to the nearest
gene, the gene name, and the gene type.}
\usage{
calcNearestGenes(query, annotations, removeUnknowns = TRUE)
}
\arguments{
\item{query}{A GRanges or GRangesList object with query sets}

\item{annotations}{A GRanges or GRangesList object with annotation sets}

\item{removeUnknowns}{Boolean value to indicate if you'd like to remove unannotaed ranges in the query.}
}
\value{
A data table that contains observations for each genomic region
        and the associated aforementioned annotations.
}
\description{
Find overlapping annotations for a given query and annotation
set. Annotate the query based on the distance to the nearest
gene, the gene name, and the gene type.
}
\examples{
queryFile = system.file("extdata", "vistaEnhancers.bed.gz", package="GenomicDistributions")
query = rtracklayer::import(queryFile)
data(TSS_hg19)

queryAnnotated = calcNearestGenes(query, TSS_hg19)
}
