% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor-distances.R
\name{calcNeighborDist}
\alias{calcNeighborDist}
\title{Group regions from the same chromosome together and
calculate the distances between neighboring regions. 
Distances are then lumped into a numeric vector.}
\usage{
calcNeighborDist(query)
}
\arguments{
\item{query}{A GRanges or GRangesList object.}
}
\value{
A numeric vector or list with different vectors containing the
 distances within neighboring regions.
}
\description{
Group regions from the same chromosome together and
calculate the distances between neighboring regions. 
Distances are then lumped into a numeric vector.
}
\examples{
dist = calcNeighborDist(vistaEnhancers)
}
