% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-plots.R
\name{calcPartitionPercents}
\alias{calcPartitionPercents}
\title{Calculate the percentage overlap of a list of GRanges object against a 
list of known partitions.}
\usage{
calcPartitionPercents(listGR, partitionList, backgroundGR = NULL)
}
\arguments{
\item{listGR}{A list of GRanges objects. Produced by 
\code{calcPartitionPercents}}

\item{partitionList}{A list with GRanges objects corresponding to
the different partitions.}

\item{backgroundGR}{A GRanges object to remove from listGR overlap 
as background.}
}
\value{
A named list of the frequency of input regions against provided
    partitions.
}
\description{
Calculate the percentage overlap of a list of GRanges object against a 
list of known partitions.
}
\examples{
partitionList = genomePartitionList(geneModels_hg19$genesGR, 
                                    geneModels_hg19$exonsGR,
                                    geneModels_hg19$threeUTRGR, 
                                    geneModels_hg19$fiveUTRGR)
partPerc = calcPartitionPercents(vistaEnhancers, partitionList)
}
