% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-plots.R
\name{calcPartitionsRef}
\alias{calcPartitionsRef}
\title{Calculates the distribution of overlaps for a query set to a reference 
assembly}
\usage{
calcPartitionsRef(query, refAssembly, metric = NULL, backgroundGR = NULL)
}
\arguments{
\item{query}{A GenomicRanges or GenomicRangesList object with query regions}

\item{refAssembly}{A character vector specifying the reference genome
assembly (*e.g.* 'hg19'). This will be used to grab chromosome sizes 
with \code{getTSSs}.}
}
\value{
A data.frame indicating the number of query region overlaps in   
    several genomic partitions.
}
\description{
This function is a wrapper for \code{calcPartitions} that uses built-in
partitions for a given reference genome assembly.
}
\examples{
f = system.file("extdata", "vistaEnhancers.bed.gz",
    package="GenomicDistributions")
query = rtracklayer::import(f)
calcPartitionsRef(query, "hg19")
}
