% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-plots.R
\name{calcPropPartitionsRef}
\alias{calcPropPartitionsRef}
\title{Calculates the distribution of overlaps (in base pairs) between 
query and arbitrary genomic partitions. The function uses built-in
partitions for a given reference genome assembly.}
\usage{
calcPropPartitionsRef(query, refAssembly)
}
\arguments{
\item{query}{A GenomicRanges or GenomicRangesList object with query regions}

\item{refAssembly}{A character vector specifying the reference genome
assembly (*e.g.* 'hg19'). This will be used to grab genome 
annotations with \code{getGeneModels}}
}
\value{
A data.frame giving total overlap [bp] of a GRanges object to a
    partition from a previously provided partitionList.
}
\description{
Calculates the distribution of overlaps (in base pairs) between 
query and arbitrary genomic partitions. The function uses built-in
partitions for a given reference genome assembly.
}
\examples{
calcPropPartitionsRef(vistaEnhancers, "hg19")
}
