% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cellTypeMetadata}
\alias{cellTypeMetadata}
\title{Cell type metadata}
\format{A data table with tissue and group assigned to cell type}
\description{
A dataset containing metadat for coloring cell specific open chromatin 
signal plot
}
