% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cellTypeMetadata}
\alias{cellTypeMetadata}
\title{Table the maps cell types to tissues and groups}
\format{
data.table with 3 columns (cellType, tissue and group) and 74 rows (one per cellType)
}
\source{
self-curated dataset
}
\usage{
data(cellTypeMetadata)
}
\description{
Table the maps cell types to tissues and groups
}
\keyword{datasets}
