% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{geneModels_hg19}
\alias{geneModels_hg19}
\title{hg38 gene models}
\format{
A list of two GRanges objects, with genes and exons locations
}
\source{
EnsDb.Hsapiens.v75 package
}
\description{
A dataset containing gene models for Homo Sapiens hg38 genome assembly.
}
