% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-plots.R
\name{genomePartitionList}
\alias{genomePartitionList}
\title{Create a basic genome partition list of genes, exons, introns, and 
intergenic}
\usage{
genomePartitionList(genesGR, exonsGR, threeUTRGR, fiveUTRGR)
}
\arguments{
\item{genesGR}{a GRanges object of gene coordinates}

\item{exonsGR}{a GRanges object of exons coordinates}
}
\value{
A list of GRanges objects, each corresponding to a partition of the 
    genome. Partitions include proximal and core promoters, exons and 
    introns.
}
\description{
Given GRanges for genes, and a GRanges for exons, returns a list of GRanges
corresponding to various breakdown of the genome, based on the given
annotations; it gives you proximal and core promoters, exons, and introns.
}
\details{
To be used as a partitionList for \code{calcPartitions}.
}
\examples{
partitionList = genomePartitionList(geneModels_hg19$genesGR, 
                                    geneModels_hg19$exonsGR,
                                    geneModels_hg19$threeUTRGR, 
                                    geneModels_hg19$fiveUTRGR)
}
