% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{GenomicDistributions}
\alias{GenomicDistributions}
\alias{GenomicDistributions-package}
\title{If you have a set of genomic ranges, the GenomicDistributions R package can
help you with some simple visualizations. Currently, it can produce two kinds
of plots: First, the chromosome distribution plot, which visualizes how your
regions are distributed over chromosomes; and second, the feature
distribution plot, which visualizes how your regions are distributed relative
to a feature of interest, like Transcription Start Sites (TSSs).}
\description{
If you have a set of genomic ranges, the GenomicDistributions R package can
help you with some simple visualizations. Currently, it can produce two kinds
of plots: First, the chromosome distribution plot, which visualizes how your
regions are distributed over chromosomes; and second, the feature
distribution plot, which visualizes how your regions are distributed relative
to a feature of interest, like Transcription Start Sites (TSSs).
}
\references{
\url{http://github.com/nsheff}
}
\author{
Nathan C. Sheffield
}
