% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{getChromSizes}
\alias{getChromSizes}
\title{Returns built-in chrom sizes for a given reference assembly}
\usage{
getChromSizes(refAssembly)
}
\arguments{
\item{refAssembly}{A string identifier for the reference assembly}
}
\value{
A vector with the chromosome sizes corresponding to a 
specific genome assembly.
}
\description{
Returns built-in chrom sizes for a given reference assembly
}
\examples{
getChromSizes("hg19")
}
