% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{labelCuts}
\alias{labelCuts}
\title{Creates labels based on a discretization definition.
If you are building a histogram of binned values, you want to have labels for
your bins that correspond to the ranges you used to bin. This function takes
the breakpoints that define your bins and produces nice-looking labels for
your histogram plot.}
\usage{
labelCuts(breakPoints, digits = 1, collapse = "-", infBins = FALSE)
}
\arguments{
\item{breakPoints}{The exact values you want as boundaries for your bins}

\item{digits}{Number of digits to cut the labels off.}

\item{collapse}{Character to separate the labels}

\item{infBins}{use >/< as labels on the edge bins}
}
\description{
\code{labelCuts} will take a cut group, (e.g., a quantile division of 
some signal), and give you clean labels (similar to the cut method).
}
\examples{
labelCuts(seq(0,100,by=20))
}
