% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadEnsDb}
\alias{loadEnsDb}
\title{Load selected EnsDb library}
\usage{
loadEnsDb(genomeBuild)
}
\arguments{
\item{genomeBuild}{string, genome identifier}
}
\value{
loaded library
}
\description{
Load selected EnsDb library
}
\examples{
\dontrun{
loadEnsDb("hg19")
}
}
