% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-plots.R
\name{plotcFRiF}
\alias{plotcFRiF}
\title{Plot cumulative Fraction of Reads in Features (cFRiF)}
\usage{
plotcFRiF(query, num_reads, feature_names = NA)
}
\arguments{
\item{query}{A GenomicRanges or GenomicRangesList object with query regions.
The name of each GRanges object is the default feature name.
Each GRanges object must include a counts metadata column.}

\item{num_reads}{Numeric value representing the total number of aligned
reads/bases}

\item{feature_names}{An optional character vector of feature names, in the 
same order as the GenomicRanges or GenomicRangesList 
object.}
}
\description{
This function plots the cumulative fraction of reads in a set of features
}
\examples{
data("promoter")
data("promoter_flanking")
data("exon")
data("intron")
data("utr3")
data("utr5")
plotcFRiF(query = c("promoter", "promoter_flanking", "exon",
                   "intron", "utr3", "utr5"),
          num_reads=87520,
          names = c("promoter", "promoter_flanking", "exon",
                   "intron", "utr3", "utr5")
         )
}
\keyword{cFRiF}
