% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom-plots.R
\name{plotChromBins}
\alias{plotChromBins}
\title{Plot distribution over chromosomes}
\usage{
plotChromBins(genomeAggregate, binCount = 10000,
  plotTitle = "Distribution over chromosomes")
}
\arguments{
\item{genomeAggregate}{The output from the genomicDistribution function}

\item{binCount}{Number of bins (should match the call to
\code{genomicDistribution})}

\item{plotTitle}{Title for plot.}
}
\description{
Plots result from \code{genomicDistribution} calculation
}
