% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-plots.R
\name{plotFeatureDist}
\alias{plotFeatureDist}
\title{Plots a histogram of distances to genomic features}
\usage{
plotFeatureDist(dists, featureName = "features")
}
\arguments{
\item{dists}{Results from \code{featureDistribution}}

\item{featureName}{Character vector for plot labels (optional).}
}
\description{
Given the results from \code{featureDistribution}, plots a histogram of
distances surrounding the features of interest
}
