% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-distance-plots.R,
%   R/feature-overlap-plots.R
\name{plotFRiF}
\alias{plotFRiF}
\title{Plot Fraction of Reads in Features (FRiF)}
\usage{
plotFRiF(dists, feature_names = NA)

plotFRiF(dists, feature_names = NA)
}
\arguments{
\item{dists}{Results from \code{calcExpectedDistribution}}

\item{feature_names}{An optional character vector of feature names, in the 
same order as the dists object.}

\item{query}{A GenomicRanges or GenomicRangesList object with query regions.
The name of each GRanges object is the default feature name.
Each GRanges object must include a counts metadata column.}

\item{num_reads}{Numeric value representing the total number of aligned
reads/bases}

\item{genome_size}{Numeric value representing the size of a genome in bp}
}
\description{
This function plots the fraction of reads in a set of features

This function plots the log10 of the observed/expected distribution of
reads/bases in a set of features.
}
\examples{
data("promoter")
data("promoter_flanking")
data("exon")
data("intron")
data("utr3")
data("utr5")
plotFRiF(query = c("promoter", "promoter_flanking", "exon",
                    "intron", "utr3", "utr5"),
         num_reads=87520, genome_size = 3099922541, 
         feature_names = c("promoter", "promoter_flanking", "exon",
                           "intron", "utr3", "utr5")
         )
}
\keyword{FRiF}
