% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content-plots.R
\name{plotGCContent}
\alias{plotGCContent}
\title{Plots a density distribution of GC vectors
Give results from the \code{calcGCContent} function, this will produce a
density plot}
\usage{
plotGCContent(gcvectors)
}
\arguments{
\item{gcvector}{A vector of GC contents}
}
\description{
Plots a density distribution of GC vectors
Give results from the \code{calcGCContent} function, this will produce a
density plot
}
