% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor-distances.R
\name{plotNeighborDist}
\alias{plotNeighborDist}
\title{Plot the distances between neighboring regions.The distance in the 
x axis is log10 transformed for ease of comparison between 
different regionsets and to account for outliers.}
\usage{
plotNeighborDist(dcvec)
}
\arguments{
\item{dcvec}{A numeric vector or list with vectors containing distances 
between neighbor regions. Produced by \code{calcNeighborDist}}
}
\value{
A ggplot density object showing the distribution of
log10 transformed distances.
}
\description{
Plot the distances between neighboring regions.The distance in the 
x axis is log10 transformed for ease of comparison between 
different regionsets and to account for outliers.
}
\examples{
numVector = rnorm(400, mean=5, sd=0.1)
d = plotNeighborDist(numVector)
}
