% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qthist.R
\name{plotQTHist}
\alias{plotQTHist}
\title{Plot quantile-trimmed histogram}
\usage{
plotQTHist(
  widths,
  EndBarColor = "gray57",
  MiddleBarColor = "gray27",
  quantile = NULL,
  bins = NULL
)
}
\arguments{
\item{widths}{Results from \code{calcWidths}}

\item{EndBarColor}{Color for the quantile bars on both ends of the graph (optional)}

\item{MiddleBarColor}{Color for the bars in the middle of the graph (optional)}

\item{quantile}{Quantile of data to be contained in each end bar (optional)
Quantiles must be under .2, optimal size is under .1}

\item{bins}{The number of bins for the histogram to allocate data to. (optional)}
}
\description{
given the results from \code{calcWidth}, plots a histogram of widths
}
\details{
x-axis breaks for the frequency calculations are based on the "divisions" results from 
helper function \code{calcDivisions}
}
