% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildReferenceData.R
\name{retrieveFile}
\alias{retrieveFile}
\title{Read local or remote file}
\usage{
retrieveFile(source, destDir = NULL)
}
\arguments{
\item{source}{a string that is either a path to a local or remote GTF}

\item{destDir}{a string that indicates the path to the directory where
the downloaded GTF file should be stored. If not provided, 
a temporary directory will be used.}
}
\value{
data.frame retrieved file path
}
\description{
Read local or remote file
}
\examples{
CElegansGtfCropped = system.file("extdata", "C_elegans_cropped_example.gtf.gz", package="GenomicDistributions")
CElegansGtf = retrieveFile(CElegansGtfCropped)
}
