% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition-plots.R
\name{setLabels}
\alias{setLabels}
\title{Internal helper function for \code{plotCumulativePartitions}}
\usage{
setLabels(assignedPartitions)
}
\arguments{
\item{assignedPartitions}{Results from \code{calcCumulativePartitions}}
}
\description{
Internal helper function for \code{plotCumulativePartitions}
}
