% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-distance-plots.R,
%   R/feature-overlap-plots.R
\name{setLabels}
\alias{setLabels}
\title{Internal helper function for \code{plotcFRiF} and \code{plotFRiF}}
\usage{
setLabels(frif)

setLabels(frif)
}
\arguments{
\item{frif}{Results from \code{calcFRiF}}

\item{query}{A GRanges object}
}
\description{
Internal helper function for \code{plotcFRiF} and \code{plotFRiF}

Internal helper function for \code{plotcFRiF} and 
\code{calcExpectedDistribution}
}
