% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-overlap-plots.R, R/partition-plots.R
\name{setLabels}
\alias{setLabels}
\title{Internal helper function for \code{plotcFRiF} and 
\code{calcExpectedDistribution}}
\usage{
setLabels(assignedPartitions)

setLabels(assignedPartitions)
}
\arguments{
\item{assignedPartitions}{Results from \code{calcCumulativePartitions}}

\item{frif}{Results from \code{calcFRiF}}
}
\description{
Internal helper function for \code{plotcFRiF} and 
\code{calcExpectedDistribution}

Internal helper function for \code{plotCumulativePartitions}
}
