% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{theme_blank_facet_label}
\alias{theme_blank_facet_label}
\title{Clear ggplot face label.}
\usage{
theme_blank_facet_label()
}
\value{
A ggplot theme
}
\description{
Usually ggplot2 facets are labeled with boxes surrounding the label. This
function removes the box, so it's a simple label for each facet.
}
