% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{TSS_hg19}
\alias{TSS_hg19}
\title{hg19 TSS locations}
\format{
A named vectors of lengths with one item per chromosome
}
\source{
EnsDb.Hsapiens.v75 package
}
\description{
A dataset containing chromosome sizes for Homo Sapiens hg38 genome assembly
}
